refactor: simplify dialog
